Bugfix: check result is incorrect.
During running of the CIS checks, there are
checks which give incorrect results, that is FAIL when
expected to be PASS and vice-versa.
The reason is that execute() function does string parsing
incorrectly which results in false positive/negative results.
Specifically the strings.Contains() is used which can mistakenly
find prefix of the matched flag and not the flag itself.
Another reason is that there are cases when execute()
just checks that flag is not present when it is expected to do
more careful checking. For example: 2.5. The check returns FAIL when
actually docker does not have aufs storage driver set.

Fixed by changing execute() flow using correct regex to fetch
the flag value.